+++ store-service/src/main/java/com/example/cloudfour/storeservice/domain/store/service/query/StoreQueryService.java
@@ -3,7 +3,7 @@ package com.example.cloudfour.storeservice.domain.store.service.query;
 import com.example.cloudfour.modulecommon.dto.CurrentUser;
 import com.example.cloudfour.storeservice.domain.collection.document.StoreDocument;
 import com.example.cloudfour.storeservice.domain.collection.repository.query.StoreSearchRepository;
-import com.example.cloudfour.storeservice.domain.commondto.RegionResponseDTO;
+import com.example.cloudfour.storeservice.domain.common.RegionResponseDTO;
 import com.example.cloudfour.storeservice.domain.region.exception.RegionErrorCode;
 import com.example.cloudfour.storeservice.domain.region.exception.RegionException;
 import com.example.cloudfour.storeservice.domain.store.converter.StoreConverter;
